requirements.txt
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.7.4.post0
anyio==2.2.0
appnope==0.1.2
arcgis==1.8.5.post3
argcomplete==2.0.0
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.5.5
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
branca==0.4.2
catkin-pkg==0.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colcon-argcomplete==0.3.3
colcon-bash==0.4.2
colcon-cd==0.1.1
colcon-cmake==0.2.27
colcon-common-extensions==0.3.0
colcon-core==0.10.0
colcon-defaults==0.2.6
colcon-devtools==0.2.3
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.14
colcon-output==0.2.12
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.2.4
colcon-pkg-config==0.1.0
colcon-powershell==0.3.7
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.1
colcon-ros==0.3.23
colcon-test-result==0.3.8
colcon-zsh==0.4.0
colorama==0.4.4
colored==1.4.2
contourpy==1.0.5
coverage==6.4.4
cryptography==38.0.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.12
deprecation==2.1.0
discord.py==1.7.2
distlib==0.3.6
distro==1.7.0
Django==3.2.4
dlib==19.24.0
docutils==0.19
emoji==1.2.0
empy==3.3.4
entrypoints==0.3
Fiona==1.8.18
flake8==5.0.4
flake8-blind-except==0.1.1
flake8-builtins==1.5.3
flake8-class-newline==1.6.0
flake8-comprehensions==3.10.0
flake8-deprecated==1.3
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-quotes==3.3.1
Flask==2.0.1
folium==0.12.1
fonttools==4.37.3
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopandas==0.9.0
geopy==2.1.0
gis==0.2.1
h5py==3.2.1
haversine==2.3.0
html5lib==1.1
idna==2.10
image==1.5.33
ImageFilter==0.1.0
imageio==2.9.0
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.2
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
Keras==2.4.3
keyring==23.0.1
kiwisolver==1.3.1
lark==1.1.1
lazy-object-proxy==1.6.0
lerc==0.1.0
lesscpy==0.14.0
line-bot-sdk==2.0.1
love==0.0.3
lxml==4.9.1
MarkupSafe==2.0.1
matplotlib==3.6.0
mccabe==0.7.0
mistune==0.8.4
mock==4.0.3
multidict==5.1.0
munch==2.5.0
mypy==0.931
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.11.0
networkx==2.5.1
nose==1.3.7
notebook==6.3.0
ntlm-auth==1.5.0
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.6.0.66
overprint==0.0.2
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkgtools==0.7.3
pluggy==0.13.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psutil==5.9.2
ptyprocess==0.7.0
py==1.10.0
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.20
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.5.0
pygame==2.0.1
PyGithub==1.55
Pygments==2.8.1
pygraphviz==1.10
PyJWT==2.1.0
pylint==2.8.0
pylist==1.4.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.17.3
pyshp==2.1.3
pytest==6.2.4
pytest-cov==3.0.0
pytest-mock==3.8.2
pytest-repeat==0.9.1
pytest-rerunfailures==10.0
python-certifi-win32==1.6
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
ratelim==0.1.6
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rosdep==0.22.1
rosdistro==0.9.0
rospkg==1.4.0
scikit-image==0.18.1
scipy==1.6.2
selenium==3.141.0
Send2Trash==1.5.0
setuptools-scm==6.0.1
Shapely==1.7.1
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.9.3
testpath==0.4.4
tifffile==2021.4.8
toml==0.10.2
tomli==2.0.1
tomlkit==0.7.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.0.5
TravisPy==0.3.5
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
vcstool==0.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
withlog==0.0.3
wrapt==1.12.1
wxPython==4.1.1
yarl==1.6.3
zipp==3.4.1
